Implement Fallback Identifier
This patch allows users to specify a fallback identifier in the
configuration. This identifier is used if no identifier is specified via
URL parameter.

This functionality is mostly useful for testing, although it can also
be useful for production, since you can now specify a valid fallback for
deployments and users/testers do not have to remember a specific URL
parameter or identifier.

For example, you can now run `npm run dev` and the URL printed in the
command line will now actually work.