Improve logic of selecting the initial MP4 quality
Before this commit, the highest quality was always chosen. That's often
not a good idea for multiple reasons:
- Increases server bandwidth
- Makes video load slower
- Eat lots of mobile bandwidth (if on mobile)
- For mobile devices, drains the battery more quickly or even completely
  overwhelms the devices.

I'm a video quality snob and even I don't mind 720p videos on my iPhone
(the YouTube default), despite having a screen resolution of 2340×1080.

The logic in this commit is simply: take the largest video that
completely fits in the (potentially rotated) screen. For mobile phones,
use 1600x900 as screen resolution for the described check.