Add support for changing quality to MP4 videos
Fixes #342

I based this on db5dfce1 [1] but had to add additional changes. A few
notes.

For one, it is not sufficient to pause & resume each video on its own
when changing the quality. Everything has to be stopped, all videos
have to be loaded in the new quality settings, and then all have to be
resumed. That's why changes in `StreamProvider` were necessary.

I also changed that all videos are loaded simultaneously instead of
serially, like before. This should make quality switching a bit faster.

Another change I had to do is move the `autoplay` attribute thingy. I
first thought that autoplay couldn't possible be important if it is
stopped immediately anyway. But I too found weird behaviors in Chrome
and FF when first playing the video. The time sync was triggered a few
times, making one of the two videos jump a few times. I am not 100%
sure why that happens, but with the autoplay it is indeed fixed.

However, the `.pause()` in `waitForLoaded` posed a problem for quality
changing. It always confused the "playing state" after changing the
quality, as some videos got randomly paused. I solved this by just
moving the autoplay attribute to `loadStreamData` and also resetting it
there.

One final change is that in `loadStreamData`, `waitForLoad` is not used
anymore, but the `canplay` event is used, which is more appropriate for
that use case.

[1]: https://github.com/miesgre/opencast/commit/db5dfce1fb66d8824d3cb0d801e0e74bf3749dd0