Fix bug where mp4 autoplay videos weren't paused until all streams loaded
When starting a  dual stream video, it could regularly happen that the
user already heard the audio from one stream, but the loading spinner
was still shown. That's because MP4 streams are included with a <video>
tag that has `autoplay`. This is unfortunately required for some
browsers to properly load. Paella needs to immediately pause the video
when it loaded. It is only supposed to start playing once all streams
have loaded and a "play" command is coming from the StreamProvider.

This behavior got broken in bc37279f4: a new branch was added to
waitForLoaded that could lead to `resolve()` being called, but there the
`pause` was forgotten. I couldn't just add the `pause` as waitForLoaded
is used in many places, and we don't always want to pause. So I added
a boolean parameter to only do it when called from `loadStream`.